Move rankOrStatue and VoteWeight to shared
