+++ src/components/shared/voteWeight/index.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import LiskAmount from '../../../shared/liskAmount';
-import { tokenMap } from '../../../../constants/tokens';
+import LiskAmount from '../liskAmount';
+import { tokenMap } from '../../../constants/tokens';
 /**
  * This component acts as an adapter for diversions in consecutive versions of API
  * @param {Object} data The delegate information
  */
-const RankOrStatus = ({ data }) => {
-  const apiVersion = useSelector(state => state.network.apiVersion);
+const VoteWeight = ({ data }) => {
+  const apiVersion = useSelector(state => state.network.networks.LSK.apiVersion);
 
   return (
     <strong>
       <LiskAmount
-        val={apiVersion === '2.x' ? data.vote : data.voteWeight}
+        val={apiVersion === '2' ? data.vote : data.voteWeight}
         roundTo={0}
         token={tokenMap.LSK.key}
       />
@@ -20,4 +20,4 @@ const RankOrStatus = ({ data }) => {
   );
 };
 
-export default RankOrStatus;
+export default VoteWeight;
